Unified to a single Header for Landing page and App section
